requirements/local.txt
-r base.txt

coverage==3.7.1
django-debug-toolbar==1.2
honcho==0.5.0
django-extensions
